Split up `tracks.ts` code; moved functions to various utils files
