+++ typescript/packages/well-log-viewer/src/WellLogViewer_performance.test.tsx
@@ -7,7 +7,7 @@ import React, { Profiler } from "react";
 import WellLogViewer from "./WellLogViewer";
 import logTimes, { obj } from "./test/performanceMetrics";
 import { axisMnemos, axisTitles } from "./utils/axes";
-import type { ColorMapFunction } from "./components/ColorMapFunction";
+import type { ColorMapFunction } from "./utils/color-function";
 
 // TODO: Fix this the next time the file is edited.
 // eslint-disable-next-line @typescript-eslint/no-require-imports
