+++ typescript/packages/well-log-viewer/src/utils/generateColor.ts
@@ -1,0 +1,14 @@
+const __colors = [
+    "red",
+    "blue",
+    "orange",
+    "green",
+    "red",
+    "magenta",
+    "gray",
+    "brown",
+];
+let __iPlotColor = 0;
+export function generateColor(): string {
+    return __colors[__iPlotColor++ % __colors.length];
+}
